feat: implementar contexto EstanteContext com persistência em LocalStorage
- Criado o arquivo EstanteContext.js para gerenciar o estado global das estantes (estouLendo, jaLi, queroLer) usando Context API.
- Adicionada a funcionalidade de persistência do estado no LocalStorage, garantindo que as estantes sejam mantidas após recarregar a página.
- Atualizado o arquivo App.js para envolver os componentes com o EstanteProvider.
- Modificada a ListaLivrosView para utilizar o hook useEstante, garantindo a integração com o contexto e o LocalStorage.
